[CI/Build] Split test_int8_config into CPU config tests and CUDA/NPU smoke tests
Separate test_int8_config.py into two files aligned with codebase conventions:
- test_int8_config.py (core_model, cpu): pure config/factory unit tests using mocks
- test_int8_smoke.py (core_model, cuda, L4): real hardware smoke tests with
  @cuda_available and @npu_available skipif guards
